feat(api-types): move request DTOs into the crate
Relocate request/query types (SignUp, SignIn, NewInvoice, SendPayment,
CreateApiKey, NewBtcAddress, Register/UpdateLnAddress, RegisterWallet,
LNUrlp + NostrNIP05 query params) into a `request` module. The dtos/
modules re-export them, so handler paths are unchanged; response DTOs
stay put for stage 4.
