+++ src/application/dtos/wallet.rs
@@ -1,5 +1,5 @@
 use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use utoipa::ToSchema;
 use uuid::Uuid;
 
@@ -10,12 +10,7 @@ use crate::domains::{
 
 use super::{InvoiceResponse, PaymentResponse};
 
-/// Register Wallet Request
-#[derive(Debug, Deserialize, Clone, ToSchema)]
-pub struct RegisterWalletRequest {
-    /// User ID. Should ideally be registered in your Auth provider.
-    pub user_id: String,
-}
+pub use swissknife_api_types::RegisterWalletRequest;
 
 #[derive(Serialize, ToSchema)]
 pub struct WalletResponse {
